reformatted data ingestion
